Moved Vram Measuring to renderer added peters proccesMemMeasuring to utilz
