+++ SluaghEngine/Utilz/Utilz.vcxproj
@@ -152,7 +152,6 @@
     <ClInclude Include="..\includes\Utilz\IConsoleBackend.h" />
     <ClInclude Include="..\includes\Utilz\INIReader.h" />
     <ClInclude Include="..\includes\Utilz\Memory.h" />
-    <ClInclude Include="..\includes\Utilz\MemoryMeasuring.h" />
     <ClInclude Include="..\includes\Utilz\PathParsing.h" />
     <ClInclude Include="..\includes\Utilz\RunTimeStringHasher.h" />
     <ClInclude Include="..\includes\Utilz\StackAllocator.h" />
@@ -165,7 +164,6 @@
     <ClCompile Include="Console.cpp" />
     <ClCompile Include="DoubleEdgeStackAlloc.cpp" />
     <ClCompile Include="INIReader.cpp" />
-    <ClCompile Include="MemoryMeasuring.cpp" />
     <ClCompile Include="Source.cpp" />
     <ClCompile Include="StackAllocator.cpp" />
     <ClCompile Include="Tools.cpp" />
