+++ SluaghEngine/Utilz/Utilz.vcxproj.filters
@@ -48,9 +48,6 @@
     <ClInclude Include="..\includes\Utilz\GUID.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\includes\Utilz\MemoryMeasuring.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="..\includes\Utilz\DoubleEdgeStackAlloc.h">
       <Filter>Header Files</Filter>
     </ClInclude>
@@ -84,9 +81,6 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="StackAllocator.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="MemoryMeasuring.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="DoubleEdgeStackAlloc.cpp">
