+++ SluaghEngine/includes/Graphics/IRenderer.h
@@ -11,6 +11,7 @@
 #include "AnimationStructs.h"
 #include "FileHeaders.h"
 
+
 #if defined DLL_EXPORT_RENDERER
 #define DECLDIR_R __declspec(dllexport)
 #else
@@ -278,6 +279,8 @@ namespace SE
 			virtual int CreateSkeleton(JointAttributes* jointData, size_t nrOfJoints) = 0;
 
 			virtual int CreateAnimation(DirectX::XMFLOAT4X4* matrices, size_t nrOfKeyframes, size_t nrOfJoints, size_t skeletonIndex) = 0;
+
+			virtual size_t GetVRam() = 0;
 		protected:
 			IRenderer() {};
 			IRenderer(const IRenderer& other) = delete;
